refactor: 예외 데코레이터 추가
