+++ app/slack/contents.py
@@ -1,13 +1,13 @@
 import ast
 import re
 from typing import Any
+from app.exception import BotException
 
-from app.logging import logger
 from app import models
 from app.config import ANIMAL_TYPE, PASS_VIEW, SUBMIT_VIEW
 from app.services import user_content_service
 from app.logging import event_log
-from app.slack.exception_handler import exception_handler_decorator
+from app.exception_handler import exception_handler_decorator
 
 
 @exception_handler_decorator
@@ -18,6 +18,7 @@ async def submit_command(ack, body, say, client, user_id: str) -> None:
     await user_content_service.open_submit_modal(body, client, SUBMIT_VIEW)
 
 
+@exception_handler_decorator
 async def submit_view(ack, body, client, view, say, user_id: str) -> None:
     event_log(user_id, event="글 제출 완료")
     await ack()
@@ -71,9 +72,10 @@ async def submit_view(ack, body, client, view, say, user_id: str) -> None:
         )
     except Exception as e:
         message = f"{user.name}({user.channel_name}) 님의 제출이 실패하였습니다. {str(e)}"
-        logger.error(message)  # TODO: 디스코드 알림 보내기
+        raise BotException(message)
 
 
+@exception_handler_decorator
 async def open_intro_modal(ack, body, client, view, user_id: str) -> None:
     event_log(user_id, event="다른 유저의 자기소개 확인")
     await ack()
@@ -104,6 +106,7 @@ async def open_intro_modal(ack, body, client, view, user_id: str) -> None:
     )
 
 
+@exception_handler_decorator
 async def contents_modal(ack, body, client, view, user_id: str) -> None:
     event_log(user_id, event="다른 유저의 제출한 글 목록 확인")
     await ack()
@@ -122,6 +125,7 @@ async def contents_modal(ack, body, client, view, user_id: str) -> None:
     )
 
 
+@exception_handler_decorator
 async def bookmark_modal(ack, body, client, view, user_id: str) -> None:
     event_log(user_id, event="북마크 저장 시작")
     await ack()
@@ -199,6 +203,7 @@ def get_bookmark_view(
     return view
 
 
+@exception_handler_decorator
 async def bookmark_view(ack, body, client, view, say, user_id: str) -> None:
     event_log(user_id, event="북마크 저장 완료")
 
@@ -227,6 +232,7 @@ async def bookmark_view(ack, body, client, view, say, user_id: str) -> None:
     )
 
 
+@exception_handler_decorator
 async def pass_command(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="글 패스 시작")
     await ack()
@@ -234,6 +240,7 @@ async def pass_command(ack, body, say, client, user_id: str) -> None:
     await user_content_service.open_pass_modal(body, client, PASS_VIEW)
 
 
+@exception_handler_decorator
 async def pass_view(ack, body, client, view, say, user_id: str) -> None:
     event_log(user_id, event="글 패스 완료")
     await ack()
@@ -253,9 +260,10 @@ async def pass_view(ack, body, client, view, say, user_id: str) -> None:
         )
     except Exception as e:
         message = f"{user.name}({user.channel_name}) 님의 패스가 실패하였습니다. {str(e)}"
-        logger.error(message)  # TODO: 디스코드 알림 보내기
+        raise BotException(message)
 
 
+@exception_handler_decorator
 async def search_command(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="글 검색 시작")
     await ack()
@@ -263,6 +271,7 @@ async def search_command(ack, body, say, client, user_id: str) -> None:
     await user_content_service.open_search_modal(body, client)
 
 
+@exception_handler_decorator
 async def submit_search(ack, body, client, view, user_id: str):
     event_log(user_id, event="글 검색 완료")
 
@@ -338,6 +347,7 @@ def _fetch_blocks(contents: list[models.Content]) -> list[dict[str, Any]]:
     return blocks
 
 
+@exception_handler_decorator
 async def back_to_search_view(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="글 검색 다시 시작")
 
@@ -477,6 +487,7 @@ def _get_keyword(body) -> str:
     return keyword
 
 
+@exception_handler_decorator
 async def bookmark_command(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="북마크 조회")
     await ack()
@@ -567,6 +578,7 @@ def _fetch_bookmark_blocks(contents: list[models.Content]) -> list[dict[str, Any
     return blocks
 
 
+@exception_handler_decorator
 async def bookmark_search_view(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="북마크 검색 시작")
 
@@ -609,6 +621,7 @@ async def bookmark_search_view(ack, body, say, client, user_id: str) -> None:
     await ack({"response_action": "update", "view": view})
 
 
+@exception_handler_decorator
 async def open_overflow_action(ack, body, client, view, say, user_id: str) -> None:
     event_log(user_id, event="북마크 메뉴 선택")
     await ack()
@@ -647,6 +660,7 @@ async def open_overflow_action(ack, body, client, view, say, user_id: str) -> No
     )
 
 
+@exception_handler_decorator
 async def bookmark_submit_search_view(ack, body, say, client, user_id: str) -> None:
     event_log(user_id, event="북마크 검색 완료")
